Fix `estimatePrecompile` acceptance test with modularized services (#10504)
This PR fixes the failing acceptance tests in the `estimatePrecompile.feature`. The original flag `hedera.mirror.web3.evm.modularizedServices` cannot be used directly in the test module, so it was temporarily duplicated in the test module as well under `hedera.mirror.test.acceptance.web3.modularizedServices`. There is currently no common place for Spring configurations for all modules, so this is why the duplication was needed. What is more this property will be needed for a couple of more sprints and then it will be deleted from both of the modules. I also didn't see the need to define a common place for Spring properties for just 1 property that will be used in 2 modules temporarily.

A new enum was defined - `ContractMethodsModularizedServices`. It has the updated actual gas used values for the currently failing tests. In the problematic tests the `modularizedServices` flag value is compared and the correct enum value is returned.

The base `HederaTokenService` contract was updated to reflect the correct return value for `redirectForToken(...) function` and the affected contracts were recompiled and replaced as bytecode.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>