Add EntityIdSingleton for modularized (#10503)
This PR add an EntityIdSingleton implementation for getting a peek of the latest entity id.

This is needed because currently without custom implementation and db reads if we make a contract call that internally has a contract create - the contract will be assigned the first id after the last system account clashing with already existing accounts.

EntityIdSingleton - implements get method - to retrieve the first available entityId from db
EntityRepository - adds query to retrieve the first available entityId after the last system account

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>