Adapt web3 startup with modularized for non genesis startup (#10588)
Seems like during state initialization when we do migrations we have the options to pass previous version (SoftwareVersion).

Currently we are passing null as previous version which later on during the migrations is checked if null then do all genesis startup logic and validations which is pretty big operation and takes a lot of time.
This is a good approach to do on empty db like integration tests startup.

But after testing with mainnet snapshots we are getting into trouble since our db is already populated with all system accounts etc and we get into failures.

This pr assigns a previous version if db is not empty if completely empty then we should proceed with genesis logic otherwise couple of our singletons querying for record files will start failing.

Changes:
MirrorNodeState - add logic to set previous version during startup on non empty db.
tests.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>