Fix stack frames not initialized with modularized (#10619)
When running with modularized logic but traffic still has to be split to go through mono we missed to check the params modularized flag to see if we need to initialize stack frames for running mono logic and calls start to fail.

This pr adjusts the if statement where we initialize the stacked frames initialization based on the modularized flags.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>