Adapt acceptance tests and web3 to be able to force 100% traffic to mod (#10654)
This PR adds a temporary parameter Is-Modularized in the web3 header.
This is needed because there is a difference between gasUsed and some methods return results, when run against mono or modularized.

We introduced a traffic percentage which can split % of the requests agains mono and % agains modularized.
Problem is that when the web3 is deployed with a trafic split not equal to 0.0 or 1.0 acceptance tests can start failing when run against the web3 because its expects result for modularized evm but the call was processed by mono.

To fix this this new temporary Is-Modularized header forces traffic to go 100% to modularizedEvm ONLY when the evmProperties.isModularized=true.

Changes applied:

Test Module:

MirrorNodeClient - adapt web3 post request to pass the Is-Modularized when we run the tests with modularized

Web3:
ContractController - read new Is-Modularized header and force traffic to modularized if set to true.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>