Fix flakiness in historical tests by trimming total supply from getTokenInfo response in historical.feature tests (#10668)
Add a workaround for a flaky behaviour in tests that are calling getTokenInfo. A method that trims the totalSupply from the response is added.  The flakiness occurs because sometimes the response from getTokenInfo historical returns 0 for total supply instead of the real value. It is because the query for totalSupply historical depends on the db tables - `token_balance`, `token_transfers` and `account_balance` and because they update on every 15 minutes.
If the environment (e.g local node) is recently started and the db tables are still not populated with entries (the 15 mins update hasn't processed yet) the query for total supply will calculate and return 0. If there are entries, the result will be correct. 
- I could reproduce it only on local environment and only right after I started local node and the db tables were still empty. 
- Tried to reproduce it on previewnet, but couldn't. 

Another approach was to add await on the assertion at the end of the test, but this might cause a delay of 15-20 minutes in test execution which is not good.

At the end the totalSupply is not needed in those tests, because we update the name, memo and some additional parameters of the token

---------

Signed-off-by: Nikolay Nikolov <nikolay.nikolov@limechain.tech>