Add `EntityCountsSingleton` (#10689)
During the modularizedServices flow testing we found the need to implement an EntityCountsSingleton that is needed for the hedera-app dependency. This PR adds a very basic implementation which should be enough for our purposes. The value of the singleton is updated during the transaction but it is actually needed in the persist logic in hedera-app, which in web3 we don't do anyway, so we don't need to query the DB and populate real data.

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>