Implement retry and caching for system files in modularized flow (#10687)
This PR implements a throttle definition parsing mechanism with retry and caching.
If not valid definitions is found in DB we resolve to system file loader.

Mechanism
Retry and parse file data up to 10 times(Similar to RatesAndFeesLoader). Return the first valid from DB. Cache the result
In case of empty DB data - resolve to systemFileLoader -  Cache the result
In case of 10 failures - resolve to systemFileLoader - Cache the result


`SystemFileLoader` - add the retry mechanism and cache throttles similar to RatesAndFeesLoader. Method can be reused when we add retry and caching for rates and fees again for modularized
`SystemFileLoaderTest` - adds tests

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>