Adapt web3 k6 estimate tests to work with environment variables (#10653)
This PR adapts the web3 k6 estimate tests to work with environment variables. This is needed for convenience when we need to make the performance testing. The test environment variables are updated and ready to be used against the mainnet snapshot DB. _At this point they will not work directly against mainnet (valid for the estimate and the modification tests)._

2 more tests need to be fixed but this will be done in a future PR: `contractCallRedirectApprove` and `estimateDissociateTokens`

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>