Fix formatSlot function used in get contract state rest api call (#10882)
During get contract state REST API call, fomatSlot function is called if a slot query parameter is passed. For certain values ,as stated in the bug ticket #10732 ,it doesn't properly format the slot and the db query is not properly constructed and as a result no records are returned in the response.

---------

Signed-off-by: sdimitrov9 <stoyan.dimitrov@limechain.tech>
Signed-off-by: Xin Li <xin@hashgraph.com>
Co-authored-by: Xin Li <xin@hashgraph.com>