Add validation for smart contract sender accounts (#11088)
This PR enhances the validation of the sender account. Currently, we only check if it exists and is not deleted. Now we are going to check if the sender is a smart contract and fail softly with PAYER_ACCOUNT_NOT_FOUND response code as it would otherwise fail when reaching the consensus node.

All of the entity to account mapping implementation was moved from AccountReadableKVState to a new AbstractAliasedAccountReadableKVState which is now the parent of AccountReadableKVState and AliasesReadableKVState. In both these child classes we read the same entity in the DB but by different key. By caching the entity in both caches (aliases and accounts) we ensure that we will make only a single call to the DB and this way we will lighten the load on the DB.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>