Split downloaded files by date into different folders (#11024)
The mirror node is already able to read and parse files organized by folders representing days, but the organization of the files in such a way is up to the operators.

Example of already supported file organization

```shell
./data/streams/2025-04-21/recordstreams/record0.0.4/2025-04-21T23_22_12.797932000Z.rcd.gz 
./data/streams/2025-04-21/recordstreams/record0.0.4/sidecar/2025-04-21T23_22_10.184472599Z_01.rcd.gz 
```

The mirror node currently stores downloaded files in a single folder for each consensus node, and when downloading record files, signatures, and sidecars, this quickly makes file system management very difficult.

With this PR:

- With the file saving options enabled, the mirror node will separate files into different folders based on the day.
- The importer process will now stop with a graceful shutdown if the environment variable `HIERO_MIRROR_IMPORTER_DOWNLOADER_HALTONERROR` is set and there are errors saving the files (i.e. because the disk is full).


Signed-off-by: Giuseppe Bertone <giuseppe.bertone@proton.me>