Improving state key registry (#11226)
This PR makes changes to several classes to encapsulate state and singleton creation into StateKeyRegistry. A lookup method was created in said class that returns Object that is either present in state or created as default instance. The types can be ReadableKVState, ConcurrentLinkedDeque or DefaultSingleton, depending on the state definition. Changes in other classes were made to accommodate for this. Added unit and integration tests.

---------

Signed-off-by: filev94 <todor.filev@limechain.tech>