Utilize child error transactions in web3 (#11319)
This PR makes changes to utilize child transaction errors in web3 modularized flow. There are cases where we have CONTRACT_REVERT_EXECUTED with no message or detail in the parent transaction but the real reason it failed in services is present in the child transactions. Added unique statuses to childTransactionErrors list and propagate them to GenericResponse. There is a case where the parent transaction is SUCCESS but it has child transactions that failed - logging a warning in that case.

---------

Signed-off-by: filev94 <todor.filev@limechain.tech>