Acceptance tests should not encode shard and realm in EVM addresses (#11292)
This PR makes changes in the acceptance tests so that toSolidityAddress() is not used until it is fixed upstream. Instead of the latter, I am using DomainUtils.toEvmAddress(long num). Modified asAddress methods to accommodate for these changes as well as when the switch back to toSolidityAddress needs to happen, change in small amount of places should be made.

---------

Signed-off-by: filev94 <todor.filev@limechain.tech>
Signed-off-by: Jesse Nelson <jesse@hashgraph.com>
Co-authored-by: Jesse Nelson <jesse@hashgraph.com>