Adjusting gasPerSecond and gasLimitRefundPercent based on metrics (#11329)
This PR makes changes to gasPerSecond and gasLimitRefundPercent to adjust them based on metrics GasConsumed and GasLimit, taken from mainnet clusters in the time period - 05/27/2025 - 06/06/2025. Based on analysis, following observations can be made:

Gas Consumed

    Peak usage is ~67.2M on 05/30/2025 between 20:23:00 and 20:25:00 UTC+3
    95% of calls consume under ~61M.

Gas Limit

    Clients often over-request gas
    Some request billions, but only consume a fraction

Based on this, I've made changes so that the gasPerSecond is 75 Million, which is around 12% buffer above the max value. Could be a bit tight and potential change to 85 Mill or even 100 Mill is possible. Restored the gasLimitRefundPercent to its previous value of 20 to prevent over-abuse

---------

Signed-off-by: filev94 <todor.filev@limechain.tech>