Add workaround for long zero addresses to be permitted by the EVM (#11388)
Allow modularized workflow to work even if an account having an alias is referenced by its long zero (shard/realm/num) address. This will serve as a workaround for one of the breaking changes connected to the migration from mono to modularized workflow inside web3.

In order the workaround to be utilized, 2 classed needed to be overriden in the classpath. These are DispatchingEvmFrameState and ConversionUtils.

The only change in DispatchingEvmFrameState is inside isNotPriority, where based on a feature flag we return that the address is always a priority in case the feature flag is activated and if it's not then the actual check is performed.

The only change in ConversionUtils is inside accountNumberForEvmReference. If the feature flag is activated then even if an account having an alias is compared with a passed long zero address, then no NON_CANONICAL_REFERENCE_NUMBER error will be thrown. If the feature flag is disabled and the comparison has a mismatch, then the error will be thrown.

---------

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@gmail.com>