Add workaround for INSUFFICIENT_PAYER_BALANCE in modularized (#11411)
This PR add a config which when enabled will avoid INSUFFICIENT_PAYER_BALANCE errors which is a new breaking change in modularized codebase.

How it works:
Enable the overridePayerBalanceValidation flag.

Configure a minimum account balance using MirrorNodeEvmProperties.minimumAccountBalance.

When overridePayerBalanceValidation is enabled, and an account is being read from AccountReadableKVState, the following logic applies:

If the account’s actual balance in the database is less than the configured minimumAccountBalance, the system overrides the balance and sets it to the configured minimum.

This ensures the payer balance validation in hedera-app passes, as long as the minimum is set high enough.

Exception:
This override does not occur during balanceOf calls, where the real account balance must be returned.

---------

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>