Merge duplicated error log entries into one (#11434)
When a MirrorEvmTransactionException is thrown, currently there are 2 log entries for 1 transaction and this makes the investigation harder as you need to find both entries and match them. The first entry contained the request body input and the second entry contained the resulting error output.

With this PR both log entries are now merged into one. If the length of the log gets bigger than the maxPayloadLogSize property, it gets encoded as it did before for the input data. Also, the empty components in the error message are now removed and added to the log only if they are not empty.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>