Add findStorage k6 test contracts and docs (#11595)
This PR adds on more contract on mainnet that does couple of thousand storage calls to be used for k6.
The newly deployed contracts are added to the test module files.
They can be used in the already existing findStorage k6 test that we have.

Added more docs for the findStorage k6 test.

Signed-off-by: Kristiyan Selveliev <kristiyan.selveliev@limechain.tech>