Fix contract log indexes (#11535)
This PR adds a DB migration that updates the persisted contract log indexes to increment on block level and not - on transaction level as it was before. This is needed for Ethereum compatibility. For the new contract logs that will be imported a change in the importer was made as well - the record file now keeps a log index between the different processed record items and they are persisted with the correct index directly.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>