Adapt importer to handle changes introduced in HIP-1249 (#11966)
This PR adds new logic in importer related to setting gasUsed with the gasConsumed field for ContractResult in case, the RecordItems has a HAPI version equal or bigger to the HAPI version that introduces the roll out of HIP-1249. This HAPI version is dynamically adjustable with a parameter, until the exact HAPI version is set in stone. After that, this parameter can be deleted.

For RecordItems with older HAPI version the existing logic will continue to apply. This is getting the gas usage using sidecar information.

---------

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@gmail.com>