Fix "Unhandled failure for input" error (#12059)
In some rare cases when the sender of a transaction is a hollow account and we have an inner child transaction that makes a transfer from this account to another account, the first one needs to be completed. However, by calling the TransactionExecutor directly in the MN, we bypass the account completion (which happens in the HandleWorkflow on the consensus node). This PR modifies the sender hollow account in the state by setting a default key for it and this way "completing" it, so that the transaction execution can proceed as expected. After this change the transaction passed against MN and CN will be successful in both cases.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>