Return null in case token with negative num value is being read (#12124)
It seems users are passing evm address value as token addresses, which results to decoding them in a very big number that overflows Long type and produces TokenID with negative num value.

In case of such a wrong input in the calldata this PR returns null, instead of trying to decode this TokenID, the PR changes the logic to just return a null value and most likely lead to the transaction getting reverted. This will also clear out Grafana logs with unnecessary exception statements.

---------

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@gmail.com>