Add long lived cache for system accounts (#12187)
Since multiple system accounts are queried during most of the requests(payerAccount, nodeRewardAccount, stakingRewardAccount and fundingAccount) it is sensible to cache them for longer than the entity cache lifetime (2 seconds). That is why this PR introduces another cache with 5m lifetime on top of the original Entity cache that caches all entities with id< 1000 in case balanceOf operation is not initiated. (balanceCall flag is set to true when we enter CustomBalanceOperation)

---------

Signed-off-by: sdimitrov9 <stoyan.dimitrov@limechain.tech>