Enhance the contract results REST endpoint to return latest failed result (#12196)
This PR changes the /contracts/results/{transactionIdOrHash} endpoint in the following way - in case of WRONG_NONCE or DUPLICATE_TRANSACTION result they are not filtered anymore and the latest unsuccessful result is returned (if there is no successful). This way the user can know in all cases the reason why a transaction failed (when there is a contract result available).

The transaction queries are modified to reflect the change in the implementation above. The explicit filtering of both statuses is removed. Also, only the needed result is returned now by the queries. Previously, a list of results was returned and then in the contractController only the first result was used. Also, in case multiple transactions were returned, an error was thrown.

With the new implementation an error is not thrown anymore as only one result is returned by the query (the latest, matching by nonce, if provided). If a nonce is not provided, a 0 is assumed for its value. This way this matches the previous implementation where a list was returned, sorted ascending and only the first result was used (effectively nonce == 0).

To reflect the changes in this PR, multiple tests needed to be adapted to match the change in the transaction queries.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>