Rest-java k6 tests in CI (#12243)
Add rest-java k6 tests execution as part of the workflow in order to verify the end-to-end functionality. This will be helpful to catch some issues earlier in the testing cycle. The goal is not a complete coverage of all rest-java k6 tests, so the failing ones are excluded for now.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>