Update acceptance workflow to use 0.67.2-rc.2 consensus tag to fix failing acceptance tests (#12306)
Currently the acceptance workflow uses a consensus node tag that utilizes the changes from HIP-1249, having 100% gasRefundPercentage. However the mirror node importer and feature properties changed their minimum hapi minor version for these change to be 0.68. This causes inconsistencies in importer and acceptance test behaviour, since we have record files with 0.67.0 version with HIP-1249 behaviour enabled.

This PR changes the consensus version to be 0.67.2 with HIP-1249 disabled and overrides the importer to use 0.68.0 as a minimum hapi throttling version, so that acceptance tests and importer logic match.

This PR updates consensus node to a version which has HIP-1249 disabled and this should fix the failing tests.

---------

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@gmail.com>