Rest k6 tests in CI (#12242)
Add rest k6 tests execution as part of the workflow in order to verify the end-to-end functionality. This will be helpful to catch some issues earlier in the testing cycle. The goal is not a complete coverage of all rest k6 tests, so the failing ones are excluded for now. These are the historical balance tests as they are updated on longer periods of time and this will slow the workflow execution.

---------

Signed-off-by: Bilyana Gospodinova <bilyana.gospodinova14@gmail.com>