Add support for historical blocks in K6 web3 test suite (#12436)
This PR enhances the default behaviour of the K6 web3 tests to accept list of blocks which to use for test execution. For each separate block a new test run will be created. There is a new env variable property that can be set with a historical block to be used for the tests run.

---------

Signed-off-by: Ivan Kavaldzhiev <ivankavaldzhiev@gmail.com>